CAConfig variable used instead of CAType
With CAType being used as the variable, when Get-DSCConfiguration was run, the below error was received as it wasn't available. Replacing CAType with CAConfig resolves this issue and Get-DSCConfiguration is run successfully without error.
---------------------
Get-DscConfiguration : The PowerShell DSC resource  returned results that are not valid from Get-TargetResource. The CAType key is not a valid property 
in the corresponding DSC resource schema file. The results from Get-TargetResource must be in a Hashtable format. The keys in the Hashtable must be the 
same as the properties in the corresponding DSC resource schema file.
    + CategoryInfo          : InvalidResult: (MSFT_DSCLocalConfigurationManager:root/Microsoft/...gurationManager) [Get-DscConfiguration], CimException
    + FullyQualifiedErrorId : GetOperationResultInvalidResultFormat,Get-DscConfiguration
---------------------